backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
librosa==0.11.0
neucodec>=0.0.4
phonemizer==3.3.0
soundfile==0.13.1
torch==2.8.0
transformers==4.56.1
resemble-perth==1.0.1
openai==1.3.0
sentence-transformers==2.2.2
scikit-learn==1.3.0

